app/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
celery>=5.3.6
pytesseract>=0.3.10
pillow>=10.3.0
pydantic
python-dotenv
alembic==1.13.1
python-dotenv==1.0.0



